v 0.8.0 Structure update
1. Changed 'diff_abstraction' (see comments in it)
In line-node: 'sign' + 'is_updated' -> 'status'.
If status == 'updated', value = {'was': .., 'is': ..}
In diff-node: removed 'sign'. Just 'key' and 'children' in it.
2. In 'generate_diff' all data -> value of lines as it is
(so  str -> str, dicts -> dicts etc)
3. 1+2 allowed to very simplify 'generate_diff' and views: 'plain' and
'json_format', but complcated 'stylish' (because in it we need to stringify
not only nodes, but simply dicts too). However, I think it's the small fee
for common simplicity!